requirements.txt
fastapi
beautifulsoup4
readability-lxml
tldextract
python-dotenv
faiss-cpu 
sentence-transformers
transformers
python-multipart
fastapi
uvicorn
requests
lxml